Add handling of dummy user
When no users are registered in the system you can set a dummy user as active one, that has admin
priviliges. This is useful  to simplify the hole aplication. When permissions are not needed for jobs, then you do not need to register users (clients) to work with the application.